+++ src/Arcus.BackgroundJobs.AzureActiveDirectory/Arcus.BackgroundJobs.AzureActiveDirectory.csproj
@@ -15,6 +15,8 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <NoWarn>NU5125;NU5048</NoWarn>
+    <AssemblyName>Arcus.BackgroundJobs.AzureActiveDirectory</AssemblyName>
+    <RootNamespace>Arcus.BackgroundJobs.AzureActiveDirectory</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
