+++ src/Arcus.BackgroundJobs.AzureActiveDirectory/ClientSecretExpirationEventType.cs
@@ -1,9 +1,9 @@
-﻿namespace Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration
+﻿namespace Arcus.BackgroundJobs.AzureActiveDirectory
 {
     /// <summary>
     /// Represents the available event types.
     /// </summary>
-    public enum EventType
+    public enum ClientSecretExpirationEventType
     {
         /// <summary>
         /// The event type for when the client secret is about to expire.
