+++ src/Arcus.BackgroundJobs.Tests.Integration/Arcus.BackgroundJobs.Tests.Integration.csproj
@@ -25,7 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration\Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration.csproj" />
+    <ProjectReference Include="..\Arcus.BackgroundJobs.AzureActiveDirectory\Arcus.BackgroundJobs.AzureActiveDirectory.csproj" />
     <ProjectReference Include="..\Arcus.BackgroundJobs.Databricks\Arcus.BackgroundJobs.Databricks.csproj" />
     <ProjectReference Include="..\Arcus.BackgroundJobs.CloudEvents\Arcus.BackgroundJobs.CloudEvents.csproj" />
     <ProjectReference Include="..\Arcus.BackgroundJobs.KeyVault\Arcus.BackgroundJobs.KeyVault.csproj" />
