+++ src/Arcus.BackgroundJobs.Tests.Unit/Arcus.BackgroundJobs.Tests.Unit.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration\Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration.csproj" />
+    <ProjectReference Include="..\Arcus.BackgroundJobs.AzureActiveDirectory\Arcus.BackgroundJobs.AzureActiveDirectory.csproj" />
     <ProjectReference Include="..\Arcus.BackgroundJobs.Databricks\Arcus.BackgroundJobs.Databricks.csproj" />
     <ProjectReference Include="..\Arcus.BackgroundJobs.KeyVault\Arcus.BackgroundJobs.KeyVault.csproj" />
   </ItemGroup>
