+++ src/Arcus.BackgroundJobs.sln
@@ -15,7 +15,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{940B9B9D
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.BackgroundJobs.Databricks", "Arcus.BackgroundJobs.Databricks\Arcus.BackgroundJobs.Databricks.csproj", "{D38D2F5E-A151-4016-8D09-EDA6845E27C6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration", "Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration\Arcus.BackgroundJobs.AzureAD.ClientSecretExpiration.csproj", "{2E979D80-BB8C-4009-97F5-987480E96CD5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.BackgroundJobs.AzureActiveDirectory", "Arcus.BackgroundJobs.AzureActiveDirectory\Arcus.BackgroundJobs.AzureActiveDirectory.csproj", "{57F4B5A4-6DC8-4463-81A8-C12B259CDF69}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,10 +43,10 @@ Global
 		{D38D2F5E-A151-4016-8D09-EDA6845E27C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D38D2F5E-A151-4016-8D09-EDA6845E27C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D38D2F5E-A151-4016-8D09-EDA6845E27C6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2E979D80-BB8C-4009-97F5-987480E96CD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E979D80-BB8C-4009-97F5-987480E96CD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E979D80-BB8C-4009-97F5-987480E96CD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E979D80-BB8C-4009-97F5-987480E96CD5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{57F4B5A4-6DC8-4463-81A8-C12B259CDF69}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{57F4B5A4-6DC8-4463-81A8-C12B259CDF69}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{57F4B5A4-6DC8-4463-81A8-C12B259CDF69}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{57F4B5A4-6DC8-4463-81A8-C12B259CDF69}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
